separate types, replicate examples from current client- and server-side tests
